탱크 : 이동 후 공격이 멈춤 해결,  공격 데미지 수정

+++ Source/Maverick/Public/XR_LSJ/AISquad.h
@@ -16,7 +16,7 @@ struct FSquadData
 	GENERATED_BODY()
 public:
 	UPROPERTY()
-	float Hp;
+	int32 Hp;
 	UPROPERTY()
 	float AttackSpeed;
 	UPROPERTY()
